services/eval_judge/requirements.txt
openai
psycopg2-binary
